refactor: типизировать ingredient-details.tsx
