+++ src/components/ingredient-preview/ingredient-preview.tsx
@@ -1,20 +1,20 @@
 import styles from "./ingredient-preview.module.css";
 import React from "react";
 import {Text} from "../typography/text/text";
+import {TIngredient} from "../../utils/types";
 
-type TPropsIngredientPreview = {
-  image: string;
+type TPropsIngredientPreview = Pick<TIngredient, "image"> & {
   count?: number;
 };
 
 export function IngredientPreview({image, count = 0}:TPropsIngredientPreview):JSX.Element {
 
-  const text = String(count)
+  const textValue = String(count)
 
   return (
     <div style={{backgroundImage: `url(${image})`}} className={`${styles.box}`}>
       {!!count && <div className={styles.imgCardSmallFonNumber}>
-        <Text> +{text}</Text>
+        <Text> +{textValue}</Text>
       </div>}
     </div>
   )
